# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/advertisers/{advertiserId}/campaigns/{id}" : {
      "get" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "getCampaign",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCampaignResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "updateCampaign",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CampaignRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCampaignResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "deleteCampaign",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/advertisers/{advertiserId}/campaigns" : {
      "get" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "getCampaigns",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageCampaignResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "createCampaign",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CampaignRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCampaignResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/advertisers/{advertiserId}/campaigns/{id}/resume" : {
      "post" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "resumeCampaign",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/advertisers/{advertiserId}/campaigns/{id}/pause" : {
      "post" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "pauseCampaign",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/advertisers/{advertiserId}/campaigns/{id}/activate" : {
      "post" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "activateCampaign",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCampaignResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/advertisers/{advertiserId}/campaigns/wallet/topup" : {
      "post" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "topupWallet",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TopupWalletRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMapStringString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/advertisers/{advertiserId}/campaigns/{id}/performance" : {
      "get" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "getCampaignPerformance",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageCampaignPerformanceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/advertisers/{advertiserId}/campaigns/performance" : {
      "get" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "getAllCampaignPerformance",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageCampaignPerformanceResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "CampaignRequest" : {
        "required" : [ "advertiserId", "dailyBudget", "maxBid", "name", "startDate" ],
        "type" : "object",
        "properties" : {
          "advertiserId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "dailyBudget" : {
            "type" : "number"
          },
          "lifetimeBudget" : {
            "type" : "number"
          },
          "maxBid" : {
            "type" : "number"
          },
          "startDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "endDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "frequencyCap" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ApiResponseCampaignResponse" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/CampaignResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "CampaignResponse" : {
        "required" : [ "advertiserId", "dailyBudget", "frequencyCap", "id", "lifetimeBudget", "maxBid", "name", "startDate", "status", "version" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "advertiserId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "DRAFT", "SCHEDULED", "ACTIVE", "PAUSED", "COMPLETED", "ARCHIVED", "DELETED" ]
          },
          "dailyBudget" : {
            "type" : "number"
          },
          "lifetimeBudget" : {
            "type" : "number"
          },
          "maxBid" : {
            "type" : "number"
          },
          "startDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "endDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "frequencyCap" : {
            "type" : "integer",
            "format" : "int32"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "ApiResponseVoid" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "TopupWalletRequest" : {
        "required" : [ "amount" ],
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "gatewayName" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseMapStringString" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          }
        }
      },
      "ApiResponsePageCampaignResponse" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageCampaignResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageCampaignResponse" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CampaignResponse"
            }
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "PageableObject" : {
        "required" : [ "offset", "pageNumber", "pageSize", "paged", "unpaged" ],
        "type" : "object",
        "properties" : {
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          },
          "unpaged" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "paged" : {
            "type" : "boolean"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "SortObject" : {
        "required" : [ "empty", "sorted", "unsorted" ],
        "type" : "object",
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        }
      },
      "ApiResponsePageCampaignPerformanceResponse" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageCampaignPerformanceResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "CampaignPerformanceResponse" : {
        "required" : [ "advertiserId", "campaignId", "date", "id", "spend" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "advertiserId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "campaignId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "date" : {
            "type" : "string",
            "format" : "date"
          },
          "impressions" : {
            "type" : "integer",
            "format" : "int64"
          },
          "clicks" : {
            "type" : "integer",
            "format" : "int64"
          },
          "conversions" : {
            "type" : "integer",
            "format" : "int64"
          },
          "spend" : {
            "type" : "number"
          }
        }
      },
      "PageCampaignPerformanceResponse" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CampaignPerformanceResponse"
            }
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}